Update links to public and private pages
