+++ README.md
@@ -1,2 +1,5 @@
 # eventflow
+
 Smart event management platform: spaces, activities, speakers, attendees, and personalized planning.
+
+This demo uses Google Sign-In (OAuth 2.0) through Quarkus OIDC. Provide your OAuth credentials via the `GOOGLE_CLIENT_ID` and `GOOGLE_CLIENT_SECRET` environment variables.
