+++ deployment/google-oauth-secret.yaml
@@ -1,0 +1,8 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: google-oauth
+  namespace: eventflow
+stringData:
+  client-id: "YOUR_CLIENT_ID"
+  client-secret: "YOUR_CLIENT_SECRET"
