+++ quarkus-app/src/main/resources/META-INF/resources/index.html
@@ -9,8 +9,8 @@
 <body>
 <div class="container">
     <h1>Welcome</h1>
-    <p>This is a public page. <a href="/public.html">Public content</a>.</p>
-    <p>Access <a href="/private.html">private content</a> (requires login).</p>
+    <p>This is a public page. <a href="https://eventflow.opensourcesantiago.io/public.html">Public content</a>.</p>
+    <p>Access <a href="https://eventflow.opensourcesantiago.io/private.html">private content</a> (requires login).</p>
 </div>
 </body>
 </html>
