+++ quarkus-app/src/main/resources/META-INF/resources/login.html
@@ -9,11 +9,9 @@
 <body>
 <div class="container">
     <h1>Login</h1>
-    <form method="post" action="/auth/login">
-        <label>Username:<input name="username" type="text"></label><br>
-        <label>Password:<input name="password" type="password"></label><br>
-        <button type="submit">Login</button>
-    </form>
+    <p>
+        <a href="/q/login"><img alt="Sign in with Google" src="https://developers.google.com/identity/images/btn_google_signin_dark_normal_web.png"></a>
+    </p>
     <p><a href="/">Back to home</a></p>
 </div>
 </body>
