+++ quarkus-app/src/main/resources/application.properties
@@ -1,12 +1,9 @@
-# Firebase configuration
-firebase.api-key=YOUR_API_KEY
-firebase.auth-domain=YOUR_PROJECT.firebaseapp.com
-firebase.project-id=YOUR_PROJECT_ID
+# Google OAuth configuration
+quarkus.oidc.auth-server-url=https://accounts.google.com
+quarkus.oidc.application-type=web-app
+quarkus.oidc.client-id=${google.client-id}
+quarkus.oidc.credentials.secret=${google.client-secret}
+quarkus.oidc.authentication.scopes=openid email profile
 
-# OIDC configuration for Firebase
-quarkus.oidc.auth-server-url=https://securetoken.google.com/${firebase.project-id}
-quarkus.oidc.application-type=service
-quarkus.oidc.client-id=${firebase.project-id}
-quarkus.oidc.token.principal-claim=name
-quarkus.http.auth.permission.protected.paths=/protected
+quarkus.http.auth.permission.protected.paths=/private.html
 quarkus.http.auth.permission.protected.policy=authenticated
